flask_docker_compose/requirements.txt
# psycopg2
flask
psycopg2-binary